extract interface for OtherTimeSeries
